+++ changelog.md
@@ -1,3 +1,14 @@
+# dspic33a-crypto-accelerator-module v1.1.0
+### Release Highlights
+- Support for ECDH Shared Secret Generation
+- Remove device.h and crypto_config.h files from all projects
+
+### Features Added\Updated
+- Crypto Accelerator Module Demos
+- Crypto Algorithms added in this release:
+	- ECDH
+		- ECDH Shared Secret Generation
+
 # dspic33a-crypto-accelerator-module v1.0.2
 ### Release Highlights
 - Updated keywords, readmes and demo title